Add extra checks for SiGSEV on PR_void system calls (#19)
When extensions handle system calls they set the system call ID to 0
(PR_void). This is the same as what happens if a system call isn't
understood by proot. On Android unknown system calls result in a SECCOMP
failure (sending SIGSYS), rather than returning ENOSYS.

This changes the behavour of SIGSYS handing to check if the call was
handled by an extension. If the call was handled by an extention the
call should return 0, since it has been handled outside that context. If
it wasn't handled then we should return ENOSYS. This is because either
that system call doesn't exist, or we aren't allowed to call it. Either
way, since the point of proot is to adapt seccomp to other applications,
we should return ENOSYS because as far as the system is concerned that
system call doesn't exist.

Major props to nhinds for writing a bunch of go programs to help narrow
this down to the the SECCOMP filter and the handing of the SISSYS
signal.

This fixes headmelted/codebuilds#97 and CypherpunkArmory/UserLAnd#938